+++ components/ui.py
@@ -7,7 +7,7 @@ import dash_bootstrap_components as dbc
 
 from dash import dcc, html, dash_table
 from pytorch_lightning.utilities.model_summary import ModelSummary
-from lightning_pod.module import LitModel
+from lightning_pod.core.module import LitModel
 
 
 app = dash.Dash(__name__, external_stylesheets=[dbc.themes.BOOTSTRAP])
