+++ lightning_pod/core/trainer.yaml
