+++ lightning_pod/utils/new/core_seed/__init__.py
@@ -1,0 +1,1 @@
+from lightning_pod.core.trainer import main as run_trainer
