+++ lightning_pod/utils/new/core_seed/trainer_config.yaml
