--- lightning_pod/utils/new/seed/__init__.py
@@ -1,1 +1,0 @@
-from lightning_pod.trainer import main as run_trainer
